requirements.txt
aiohttp==3.8.4
python-dotenv==1.0.0
discord.py==2.2.2
Wavelink>=2.0.3